Fixes for more examples for recent cmopilers

+++ Examples/KelvinHelmholtz/SingleFluidIdealRandomSerialHDF5CheckpointRestart/main.cc
@@ -48,7 +48,7 @@ int main(int argc, char *argv[]) {
   const char* filename = "data_t0.checkpoint.hdf5";
 
   const int nOptionalSimArgs = 1;
-  std::vector<double> optionalSimArgs = {seed};
+  std::vector<double> optionalSimArgs = {static_cast<double>(seed)};
   std::vector<std::string> optionalSimArgNames = {"seed"};
 
   // Create an arg object that will contain all parameters needed by the simulation, that will be stored on the Data object.
